refactor: moved some components into folders for better organization
