+++ src/pages/signup-page.tsx
@@ -1,4 +1,4 @@
-import { SignUpForm } from "@/components/signup-form";
+import { SignUpForm } from "@/components/auth/signup-form";
 import { useAuth } from "@/context/auth-context";
 import { useLocation, Navigate } from "react-router-dom";
